fix: adapt security fixes to i128 outflow and update test assertions
- Cast ATA outflow to i128 to match the new signed fee payer outflow type
- Accept either "Insufficient token payment" or "exceeds maximum allowed"
  in the net-zero exploit test, since the i128 outflow now catches the
  exploit at the outflow limit before reaching payment validation

+++ crates/lib/src/fee/fee.rs
@@ -614,7 +614,7 @@ impl FeeConfigUtil {
         let ata_outflow =
             Self::calculate_ata_creation_outflow(fee_payer_pubkey, transaction, rpc_client, config)
                 .await?;
-        total = total.checked_add(ata_outflow).ok_or_else(|| {
+        total = total.checked_add(ata_outflow as i128).ok_or_else(|| {
             log::error!(
                 "Outflow calculation overflow in ATA accounting: sol_total={}, ata_outflow={}",
                 total,
